moved arm constants to arm folder (Rhetoric  class is boring ngl)
