+++ src/main/java/org/sciborgs1155/robot/Robot.java
@@ -14,15 +14,15 @@ import org.sciborgs1155.lib.DeferredCommand;
 import org.sciborgs1155.lib.failure.Fallible;
 import org.sciborgs1155.lib.failure.FaultBuilder;
 import org.sciborgs1155.lib.failure.HardwareFault;
-import org.sciborgs1155.robot.Constants.Elbow;
-import org.sciborgs1155.robot.Constants.Elevator;
-import org.sciborgs1155.robot.Constants.Wrist;
 import org.sciborgs1155.robot.Ports.OI;
 import org.sciborgs1155.robot.commands.Autos;
 import org.sciborgs1155.robot.subsystems.Intake;
 import org.sciborgs1155.robot.subsystems.LED;
 import org.sciborgs1155.robot.subsystems.arm.Arm;
 import org.sciborgs1155.robot.subsystems.arm.ArmState;
+import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elbow;
+import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elevator;
+import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Wrist;
 import org.sciborgs1155.robot.subsystems.arm.ArmState.GamePiece;
 import org.sciborgs1155.robot.subsystems.arm.ArmState.Goal;
 import org.sciborgs1155.robot.subsystems.drive.Drive;
