+++ src/main/java/org/sciborgs1155/robot/subsystems/arm/Arm.java
@@ -19,10 +19,10 @@ import org.sciborgs1155.lib.Trajectory;
 import org.sciborgs1155.lib.failure.Fallible;
 import org.sciborgs1155.lib.failure.FaultBuilder;
 import org.sciborgs1155.lib.failure.HardwareFault;
-import org.sciborgs1155.robot.Constants.Elbow;
-import org.sciborgs1155.robot.Constants.Elevator;
-import org.sciborgs1155.robot.Constants.Wrist;
 import org.sciborgs1155.robot.Robot;
+import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elbow;
+import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Elevator;
+import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Wrist;
 import org.sciborgs1155.robot.subsystems.arm.ArmState.GamePiece;
 import org.sciborgs1155.robot.subsystems.arm.ArmState.Goal;
 import org.sciborgs1155.robot.subsystems.arm.ElevatorIO.ElevatorConfig;
