+++ src/main/java/org/sciborgs1155/robot/subsystems/arm/RealWrist.java
@@ -1,6 +1,5 @@
 package org.sciborgs1155.robot.subsystems.arm;
 
-import static org.sciborgs1155.robot.Constants.Wrist.*;
 import static org.sciborgs1155.robot.Ports.Wrist.*;
 
 import com.revrobotics.CANSparkMax;
@@ -18,6 +17,7 @@ import org.sciborgs1155.lib.constants.MotorConfig;
 import org.sciborgs1155.lib.failure.FaultBuilder;
 import org.sciborgs1155.lib.failure.HardwareFault;
 import org.sciborgs1155.robot.Constants;
+import org.sciborgs1155.robot.subsystems.arm.ArmConstants.Wrist;  
 
 public class RealWrist implements JointIO {
 
@@ -37,13 +37,13 @@ public class RealWrist implements JointIO {
   private double lastVoltage;
 
   public RealWrist(JointConfig config) {
-    motor = MOTOR_CFG.build(MotorType.kBrushless, MOTOR);
+    motor = Wrist.MOTOR_CFG.build(MotorType.kBrushless, MOTOR);
     absolute = new DutyCycleEncoder(ABS_ENCODER);
     relative = new Encoder(RELATIVE_ENCODER[0], RELATIVE_ENCODER[1]);
 
-    absolute.setPositionOffset(ZERO_OFFSET);
-    absolute.setDistancePerRotation(CONVERSION_ABS);
-    relative.setDistancePerPulse(CONVERSION_RELATIVE);
+    absolute.setPositionOffset(Wrist.ZERO_OFFSET);
+    absolute.setDistancePerRotation(Wrist.CONVERSION_ABS);
+    relative.setDistancePerPulse(Wrist.CONVERSION_RELATIVE);
     relative.setReverseDirection(true);
 
     MotorConfig.disableFrames(motor, 4, 5, 6);
